BUG: UseImageSpacing defaults for GradientAnisotropicDiffusion
When Slicer4 was switched from ITKv3 to ITKv4, the
GradientAnisotropicDiffusionImageFilterTest started failing.  This is
because the default for the parameter "UseImageSpacing" changed from
"off" to "on" between these versions of ITK.  This patch adds an
advanced option to the CLI to enable the user to choose whether to use
image spacing, and the default is "on" in accordance with the new ITK
behavior. Tests are added that test turning this variable both on and
off, and a new regression test image is included to test the case when
the parameter is on.  With these changes, the tests should pass when
both ITKv3 and ITKv4 are used.